backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.4.0
blinker==1.9.0
brightdata-sdk==2.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
filelock==3.20.3
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
letta-client==0.1.324
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.3
pydantic-settings==2.12.0
pydantic_core==2.41.4
python-dotenv==1.1.1
requests==2.32.5
requests-file==3.0.1
sniffio==1.3.1
tldextract==5.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
